FEAT: roomService 작성자

+++ src/main/java/com/elice/homealone/room/service/RoomService.java
@@ -1,5 +1,10 @@
 package com.elice.homealone.room.service;
 
+import com.elice.homealone.global.exception.ErrorMessage;
+import com.elice.homealone.global.exception.RoomException;
+import com.elice.homealone.global.jwt.JwtTokenProvider;
+import com.elice.homealone.member.entity.Member;
+import com.elice.homealone.member.repository.MemberRepository;
 import com.elice.homealone.room.dto.RoomDto;
 import com.elice.homealone.room.dto.RoomSummaryDto;
 import com.elice.homealone.room.entity.Room;
@@ -15,14 +20,22 @@ import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Optional;
+
 @RequiredArgsConstructor
 @Slf4j
 @Service
 public class RoomService {
     private final RoomRepository roomRepository;
+    private final MemberRepository memberRepository;
+    private final JwtTokenProvider jwtTokenProvider;
+
     @Transactional
-    public RoomDto.RoomInfoDto CreateRoomPost(RoomDto roomDto){ ///회원 정의 추가해야함.
-        Room room = new Room(roomDto);
+    public RoomDto.RoomInfoDto CreateRoomPost(RoomDto roomDto,String token){ ///회원 정의 추가해야함.
+        String email = jwtTokenProvider.getEmail(token);
+        Member member = memberRepository.findByEmail(email).orElseThrow(//회원이 없을때 예외 던져주기
+                 );
+        Room room = new Room(roomDto,member);
         //HTML태그 제거
         String plainContent = Jsoup.clean(roomDto.getContent(), Safelist.none());
         room.setPlainContent(plainContent);
@@ -31,9 +44,18 @@ public class RoomService {
     }
 
     @Transactional
-    public RoomDto.RoomInfoDto EditRoomPost(Long roomId, RoomDto roomDto){
+    public RoomDto.RoomInfoDto EditRoomPost(String token,Long roomId, RoomDto roomDto){
+        if(token == null || token.isEmpty()){
+            throw new RoomException.UnauthorizedActionException(ErrorMessage.NO_JWT_TOKEN);
+        }
+        String email = jwtTokenProvider.getEmail(token);
+        Member member = memberRepository.findByEmail(email).orElseThrow(//TODO:회원이 없을때 예외 던져주기
+        );
         Room roomOriginal = roomRepository.findById(roomId)
-                .orElseThrow();
+                .orElseThrow(() ->new RoomException.RoomNotFoundException(ErrorMessage.ROOM_NOT_FOUND));
+        if(roomOriginal.getMember() != member){
+           throw new RoomException.UnauthorizedActionException(ErrorMessage.NOT_UNAUTHORIZED_ACTION);
+        }
         roomOriginal.setTitle(roomDto.getTitle());
         roomOriginal.setContent(roomDto.getContent());
         roomOriginal.setThumbnailUrl(roomDto.getThumbnailUrl());
@@ -41,9 +63,18 @@ public class RoomService {
     }
 
     @Transactional
-    public void deleteRoomPost(Long roomId){
+    public void deleteRoomPost(String token,Long roomId){
+        if(token == null || token.isEmpty()){
+            throw new RoomException.UnauthorizedActionException(ErrorMessage.NO_JWT_TOKEN);
+        }
+        String email = jwtTokenProvider.getEmail(token);
+        Member member = memberRepository.findByEmail(email).orElseThrow(//TODO:회원이 없을때 예외 던져주기
+        );
         Room roomOriginal = roomRepository.findById(roomId)
-                .orElseThrow();
+                .orElseThrow(() ->new RoomException.RoomNotFoundException(ErrorMessage.ROOM_NOT_FOUND));
+        if(roomOriginal.getMember() != member){
+            throw new RoomException.UnauthorizedActionException(ErrorMessage.NOT_UNAUTHORIZED_ACTION);
+        }
         roomRepository.delete(roomOriginal);
     }
 
@@ -76,7 +107,7 @@ public class RoomService {
     @Transactional
     public RoomDto.RoomInfoDto findByRoomId(Long roomId){
         Room room = roomRepository.findById(roomId)
-                .orElseThrow();
+                .orElseThrow(() ->new RoomException.RoomNotFoundException(ErrorMessage.ROOM_NOT_FOUND));
         room.setView(room.getView()+1);
         return RoomDto.RoomInfoDto.toRoomInfoDto(room);
     }
